SERVER-29003 Fix unsigned int warning

+++ src/mongo/db/storage/mmap_v1/mmap_windows.cpp
@@ -194,7 +194,7 @@ void* MemoryMappedFile::map(OperationContext* opCtx,
     filename[255] = 0;
     {
         size_t len = strlen(filename);
-        for (size_t i = len - 1; i >= 0; i--) {
+        for (int i = len - 1; i >= 0; i--) {
             if (filename[i] == '/' || filename[i] == '\\')
                 break;
